requirements.txt
numpy==2.2.1
pillow==10.3.0
pynput==1.7.7
psutil==6.1.1
PySide6==6.8.0.2
scipy==1.14.1
pyyaml==6.0.2
XInput-Python==0.4; sys_platform=='win32'
pyobjc-framework-Quartz==11.0; sys_platform=='darwin'
pyxhook; sys_platform!='win32' and sys_platform!='darwin'
python-xlib; sys_platform!='win32' and sys_platform!='darwin'
